finished getting all data on dom, gifs working
