+++ src/Styles/main.css
@@ -1,0 +1,73 @@
+body{
+    background-color: #36496A;
+    color: #ffff;
+}
+h1{
+background-color: #A1CDEE;
+    text-align: center;
+    margin-left: 13em;
+    margin-right: 13em;
+}
+
+fieldset {
+    flex-direction: column;
+}
+
+.formFieldset {
+    display: flex;
+    flex-direction: column;
+    font-size: 1.5rem;
+}
+
+fieldset {
+    margin: .2rem;
+    padding:.2rem;
+    border: 2px rgba(139.3,139.3,139.3,.5) solid;
+    border-radius: 6px;
+}
+.formFieldset {
+    display: flex;
+    flex-direction: column;
+    font-size: 1.5rem;
+    margin: 2 kkklem;
+}
+
+footer{
+    text-align: center;
+}
+
+main {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+	margin: 20px 0px;
+}
+
+.journal-card{
+    width: 60%;
+	/* padding-bottom: 20px; */
+	list-style-type: none;
+    background-color: #C3D5EB;
+    color: #36496A;
+    text-align: center;
+  justify-content: center;
+
+    
+}
+
+.entryList{
+   text-align: center;
+}
+
+aside{
+    width: 20%;
+   float: right;
+   overflow: scroll;
+    max-height: 400px;
+
+}
+
+.journal-image{
+    width: 300px;
+    height: 300px;
+}
