+++ src/scripts/data/dataManager.js
@@ -1,23 +1,12 @@
 export const getEntry = () => {
 
     return fetch("http://localhost:8088/entry")
-    .then(reponse => Response.json())
-    // .then(parsedResponse => {
-
-    //     return parsedResponse;
-    // })
-}
-
-export const getUsers = () => {
-    return fetch("http://localhost:8088/users")
-}
-
-const loggedInUser = {
-	id: 1,
-	name: "Autumn",
-	email: "acarte94@gmail.com"
-}
-
-export const getLoggedInUser = () => {
-	return loggedInUser;
-}
+    .then(response => response.json())
+    .then(data => {
+    const sortedByDate = data.sort(
+          (currentEntry, nextEntry) =>
+              Date.parse(currentEntry.date) - Date.parse(nextEntry.date)
+      )
+      return sortedByDate
+    })
+  };
